Fix image used in c9 standalone job with container builds
Currently when tripleo-ci-centos-9-standalone template is used with
build_container_images: true, it by default uses ubi8 image and we
always need to set the containers_base_image var.

This change adds that var in the base job so that we no longer need to
define the var in standalone jobs inheriting that template.

Closes-Bug: #1960308
Change-Id: I42092ee8633a5dfe041af8cd12f15c24b995eaf1