Move 'setup-compose-repos' pre playbook to base jobs
This patch moves setup-compose-repos.yml playbook[1] to tripleo-ci repo
to be used on its base jobs.
This playbook is responsible for configuring and enabling CentOS compose
repos based on a provided compose URL.
It has no effect if 'dependency' var is not set to 'compose-repos'.
'centos_compose_url', if omitted, will be set by default to the latest
CentOS compose (for Stream 8 or 9).
With this change, the playbook can be deleted from previous repo.
The playbook is renamed to centos-compose-repos.yml

[1] https://github.com/rdo-infra/rdo-jobs/blob/master/playbooks/dependency/setup-compose-repos.yml

Change-Id: I09cf2d74638f9c056e62fdaaa2ff41047fce925d
Signed-off-by: Douglas Viroel <dviroel@redhat.com>