Centos8 wallaby jobs teardown - standalone/scenario jobs
This removes the standalone/scenarios from running on stable
wallaby.

This includes a tidy up of the layout to avoid a lot of
repetition using yaml anchors/overrides. This is related to
the change as we had to use branches: overrides on the centos9
jobs in the layout (as well as the existing c8 ones) which
became very messy/long.

Also include reparenting for the scenario standalones as
they were inheriting from the 8 equivalent directly with
inverse branches: matches. See [1] for more context and [2]
for topic branch/test patches.

Removes obsolete c8 jobs like
tripleo-ci-centos-8-scenario010-standalone-wallaby
tripleo-ci-centos-8-scenario010-ovn-provider-standalone-wallaby
tripleo-ci-centos-8-scenario010-ovn-provider-standalone

Moves tripleo-ci-centos-8-scenario010-standalone deprecated for now

[1] http://lists.openstack.org/pipermail/openstack-discuss/2022-February/027403.html
[2] https://review.opendev.org/q/topic:c8_wallaby_teardown

Change-Id: I67bec4b9d25cf1a3d1c55f254e9c321ed979b172