Added tripleo-ci-post ansible role
playbooks/tripleo-ci/post-artifact-mult.yml and
playbooks/tripleo-ci/post-artifact-sa.yml are used in post-run
to collect zuul artifacts.

Both are defined in multinode and standalone CentOS 8 base jobs.

While doing Zuul job reparenting, We found that It cannot be used
in different jobs defined in downstream or RDO. As zuul looks for
playbook in the project where zuul jobs are defined.

In order to reuse it in downstream and RDO, we are converting it
into role and call it in starting as a part of post.yaml playbook.

It also removes it from post-run.

Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>
Change-Id: Ibae7eaa6e6e91c88ac233c8fd52550e89bf2defc