Use tripleo-buildcontainers pre playbook in RDO and Upstream
In Downstream, tripleo-buildcontainers pre playbook is different.
Due to zuul reparenting, if a job is parented from
tripleo-ci-base-containers-standard, then run and post-run
playbook will be triggered.

Since pre playbook is specific to RDO and Upstream and in
order to re-use the same we need a new parent.

It is similar to tripleo-build-containers-base-centos-8 job.

Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>
Change-Id: I6027a3ac06a4ee76316c1191bf82063eb2e9cadf