Centos8 wallaby jobs teardown - deployment image build jobs
Per the topic branch [1] and discussed at [2] this removes the
centos8 buildimage* jobs from running on stable/wallaby branch.

Some jobs are removed completely as they are specific to wallaby
and centos8 (tripleo-buildimage-ironic-python-agent-centos-8-wallaby).
The tripleo-buildimage-overcloud-hardened-uefi-full-centos-8 is moved
to deprecated for now as it is still used by DIB.

As discussed during reviews this keeps overcloud-full and IPA
for C8 wallaby for now - they may be removed by a subsequent patch.

[1] https://review.opendev.org/q/topic:c8_wallaby_teardown
[2] http://lists.openstack.org/pipermail/openstack-discuss/2022-February/027403.html

Change-Id: I04526b2c7f90ab9399fb92c86a0da216024acd30