Rename the centos-8-scenarioxxx-standalone-options jobs to generic
Removes the -centos-8 part from the -options job names as discussed
at [1]. There is an equivalent change needed in the rdo jobs that
depends-on this change at [2]. The jobs are used by both centos-8
and centos-9 and keeping the release specific part of the name is
confusing especially following the reparenting effort.

The original jobs are kept until we can merge [2] at which point
the next review in this series will remove them.

[1] https://review.opendev.org/c/openstack/tripleo-ci/+/831543/15..16/zuul.d/standalone-jobs.yaml#b798
[2] https://review.rdoproject.org/r/c/rdo-jobs/+/40842

Change-Id: I220d9b787bf1fe121ffedcaf5e1b448a3f8fbadb