Call container-registry role at required job level
In tripleo-ci-containers-rdo-upstream-pre job,
playbooks/tripleo-buildcontainers/pre.yaml is called
in pre-run where it looks for ansible-role-container-registry
role to configure docker registry.

It comes from ansible-role-container-registry project.
if it is not defined there, then the whole pre-run playbook
gets skipped.

Adding it at the job level fixes the issue.

It is only needed for CentOS-7 jobs.
Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>
Change-Id: I04779ec0ab6b6b97707f21d9d14415094bdfff20