Ensure we clean the /etc/machine-id
virt-customize (and, by extension, virt-sysprep) generate an
/etc/machine-id before anything, in order to ensure commands like
"--update" or "--install <package>" won't fail due to a missing thing.

For instance, it seems the kernel package needs the /etc/machine-id to
exist in its %post part - and it's not the only one.

You can refer to [1] in order to get some more context.

[1] https://bugzilla.redhat.com/show_bug.cgi?id=1554546

Change-Id: I185b0c86d5a8259940bdeb6d6cd1302fc3bbd7d8