Pass centos.repo from host during image build.
We are facing an issue during image build in the related bug,
because mirrors.centos.org and vexx mirrors are not in sync.

The base image which we use to build overcloud images already have
centos.repo and when proxy mirros are not updated this can cause
an issue.

This is a workaround patch to pass centos.repo(which are disabled
on host) so that same repos in the image will be overridden.

One change in behavior is at the end of image build, dib will delete
the centos.repos in the overcloud image as dib cleans up what it adds
ignoring what was already present.

Related-Bug: #1975551
Change-Id: Ia230c08817f138c2cba60076d29e1d58fde118c8