Added standalone-full-tempest-api & -scenario -options jobs
These job definitions are defined in github.com/rdo-infra/rdo-jobs
repo e.g. [1] and reduplicated in downstream jobs repo.

We can not define the -options jobs in RDO because rdo-jobs repos
are not shadowed downstream (as it brings so many job definitions,
which are hard to manage downstream). So it needs to be defined in
tripleo-ci upstream repo and inherited where ever required.

[1] https://github.com/rdo-infra/rdo-jobs/blob/3b7557a6f24f8f1776f0a70ee13eb1c32c118753/zuul.d/standalone-jobs.yaml#L245-L255
Change-Id: Ied2f16ddc3da1d2eecf49d738427cbddd0f75f8d