Run scenario004/010 if TQE roles/standalone/tasks/.* changes
This is a follow up for [1]. If there's any change to the
tasks in the TQE Ansible role standalone, then test not only
standalone 001 but also 004 and 010 since changes in those
files could also break the Ceph portion of those jobs. This
patch is depended on by [2] so we can see all Ceph jobs for
Wallaby run when that branch is moved to deployed Ceph.

[1] https://review.opendev.org/c/openstack/tripleo-ci/+/842900
[2] https://review.opendev.org/c/openstack/tripleo-quickstart-extras/+/845054

Change-Id: I3325b20945f78bd34cfac24f0e5747f43fc056a7