Fix gating repo for mixed OS content providers
This is not taking into account the mixed OS job where we
need to use registry_ip_address_distro instead of
registry_ip_address_branch.

The _branch is still default, and we override with _distro when
it is in use (i.e. when it has more than one key). The current
code causes related-bug because we end up trying to install
tripleoclient from the *8* content provider for the *9*
undercloud.

Related-Bug: 1983004
Change-Id: Ic77a58ce04153d9ee70772284a13fe1c37d9350f