Override Nameserver for vexxhost cloud
We are hitting slow dns resolution for the jobs running in vexxhoxt
cloud because 127.0.0.1 is first entry in resolv.conf but the image
we use don't have unbound service installed. Unbound service is
present in Upstream images, details in related bug.

With this patch, overriding Nameserver values used in Vexxhost cloud.

Related-Bug: #1983718
Change-Id: Ie95ed0e6b67afc28d124b2cf91fe67b57f18e654