Disable tcib_extra_config for train temporarily
tcib_extra_config features does not work in RHOS-16.2 and
the fix[1] needs to be promoted to downstream to fix it.

In order to unblock the container-build job downstream, we
are disabling this feature for train.

Note: We have not added code comment in the template change as it
breaks the CLI.

[1]. https://review.opendev.org/c/openstack/tripleo-ansible/+/857777

Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>
Change-Id: I0cde6daed826f08d8cc4b51f2b3350bd303ffdb6