Remove pingtest from tripleo
Ping test is no longer being used in the last 5 years and was replaced
by tempest. However there are still pieces of code being executed, and
replacing some files, for example the testrepository.subunit file that
tempest generate and it is being used by the ci logging scrapper to
collect data about tempest.

Change-Id: Ia2b18f88652f144d5b593c5d209a9acf2d0374c8