Add edpm base jobs
In order to add the periodic version of external compute jobs,
we need to create a seperate job which contains the reusable
job definitions. It can be reused in RDO, Downstream and
Upstream job definitions.

edpm jobs are parented from multinode cs9 distro jobs. When we
use multi parenting with the existing jobs then tags and playbook
vars are not getting overriden in multi parenting.

In order to fix that, we have added a new edpm base job which
will be shared in RDO/downstream also.

Here is the new parent structure
tripleo-ci-base-multinode-standard
tripleo-ci-base-multinode-edpm-standard
tripleo-ci-base-required-projects-edpm-upstream
tripleo-ci-base-edpm-centos-9
then actual job: tripleo-ci-centos-9-standalone-external-compute-target-host

Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>
Change-Id: I7cc3b191e9af124878f037bcf0f837d353a44f8f