Do not use --debug for image build
Since Ansible was bumped to 2.14, we've observed the container image
build process gets stuck in the middle of ansible tasks to generate
Docker/Buildah files, because of a bug[1] with ansible-runner.

This removes --debug option from the build command to avoid -vvv option
in the ansible command, to workaround the above bug.

Same workaround is added for content-provider already[2], adding same
workaround for build-containers role.

[1] https://github.com/ansible/ansible-runner/issues/1164
[2] https://review.opendev.org/c/openstack/tripleo-quickstart-extras/+/864838

Related-Bug: #1996612
Change-Id: I498c9cac7815d3d0682835d2bf943594dad2203c