Remove centos8 wallaby jobs from zuul layouts
This prevents some of the remaining c8 check and gate jobs from
running for stable/wallaby.

We are removing wallaby/c8 integration and component lines in [1]
and keeping only mixed rhel and ffu jobs with container builds as
part of the existing wallaby/centos9 line. Thus in this patch we
don't alter the wallaby/8 content provider and upgrade jobs as we
want these to continue runing. See [2] for other related patches.

[1] https://review.rdoproject.org/r/q/topic:oooci_teardown_8_wallaby
[2] https://review.opendev.org/q/topic:oooci_teardown_8_wallaby

Change-Id: Ife23e183a279417e393ee9b054e28a63b3f1298e