Update call for task include
"include" is mostly deprecated - and the "static" parameter is dead in
newer ansible.

It's better to call "include_tasks" instead, since it will dynamically
load the task. In the loop context, it will act the right way. It's also
backward compatible with older ansible versions, at least the ones we're
using in CI env.

This patch also replace the two "with_items" by "loop" parameter.

Change-Id: Id7fc8eb6c24c9671835dd217d834ae462e483da7